+++ packages/odyssey-react-mui/src/createUniqueId.ts
@@ -13,5 +13,5 @@
 // This is a random number chosen to shrink down the unique ID to an arbitrary length.
 export const uniqueIdLength = 6;
 
-export const createUniqueId = (): string =>
+export const createUniqueId = () =>
   Math.random().toString(36).slice(-uniqueIdLength);